Added: All the service, repository and controllers

+++ src/main/java/online/restadvidor/restadvidor/irepository/OpinionRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OpinionRepository extends JpaRepository<Opinion, Integer> {
+public interface OpinionRepository extends JpaRepository<Opinion, Long> {
 }
