+++ src/main/java/online/restadvidor/restadvidor/irepository/RestaurantUserRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RestaurantUserRepository extends JpaRepository<RestaurantUser, Integer> {
+public interface RestaurantUserRepository extends JpaRepository<RestaurantUser, Long> {
 }
